+++ aws/eks-cluster/nodes.tf
@@ -1,37 +1,18 @@
-resource "aws_eks_node_group" "default" {
+module "eks_managed_node_group" {
+  source = "./modules/managed-node-group"
+
   for_each = var.node_groups
 
-  cluster_name    = local.cluster_name
-  node_group_name = each.value.node_group_name
-  node_role_arn   = aws_iam_role.node.arn
-  subnet_ids      = var.private_subnet_ids
-  disk_size       = each.value.disk_size
-  instance_types  = each.value.instance_types
-  version         = var.cluster_version
+  name         = each.value.node_group_name
+  cluster_name = aws_eks_cluster.master.name
+  subnet_ids   = var.private_subnet_ids
 
-  lifecycle {
-    ignore_changes = [
-      remote_access
-    ]
-  }
+  min_size     = each.value.min_size
+  max_size     = each.value.max_size
+  desired_size = each.value.desired_size
 
-  scaling_config {
-    desired_size = each.value.desired_size
-    max_size     = each.value.max_size
-    min_size     = each.value.min_size
-  }
+  instance_types = each.value.instance_types
+  capacity_type  = "ON_DEMAND"
 
-  dynamic "taint" {
-    for_each = each.value.taint != null ? each.value.taint : []
-    content {
-      key    = taint.value.key
-      value  = taint.value.value
-      effect = taint.value.effect
-    }
-  }
-
-  depends_on = [
-    aws_iam_role_policy_attachment.node_policy_attachments,
-    aws_eks_cluster.master
-  ]
+  taint = each.value.taint
 }